moved key map to config
