+++ kr20.py
@@ -1,6 +1,5 @@
 from utils import get_item_std, get_sorted_responses
-from config import get_service_config
-from key_map import get_keyword_value
+from config import get_service_config, get_keyword_value
 
 
 def calculate_kr20(param):
